requirements.txt
torch
diffusers
transformers
accelerate
bitsandbytes
flask
Pillow
sentencepiece
